% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CodeChecking.R
\name{checkUsagePackage}
\alias{checkUsagePackage}
\title{Check all code in a package}
\usage{
checkUsagePackage(package, ignoreHiddenFunctions = TRUE,
  suppressBindingKeywords = c("ggplot2", "ffwhich", "subset.ffdf", "glm"))
}
\arguments{
\item{package}{The name of the package to check.}

\item{ignoreHiddenFunctions}{Ignore functions for which the definition cannot be retrieved?}

\item{suppressBindingKeywords}{A set of keywords that are indicative of non-standard evaluation.}
}
\description{
Check all code in a package
}
\details{
This function uses the codetools package to check the code from problems. Heuristics are used to
elimite false positives due to non-standard evaluation.
}

