% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterApply}
\alias{clusterApply}
\title{Apply a function to a list using the cluster}
\usage{
clusterApply(cluster, x, fun, ..., stopOnError = FALSE, progressBar = TRUE,
  divideFfMemory = TRUE, setFfTempDir = TRUE)
}
\arguments{
\item{cluster}{The cluster of threads to run the function.}

\item{x}{The list on which the function will be applied.}

\item{fun}{The function to apply.}

\item{...}{Additional parameters for the function.}

\item{stopOnError}{Stop when one of the threads reports an error? If FALSE, all errors will be
reported at the end.}

\item{progressBar}{Show a progress bar?}

\item{divideFfMemory}{When TRUE, the memory available for processing ff and ffdf objects will be
equally divided over the threads.}

\item{setFfTempDir}{When TRUE, the ffTempDir option will be copied to each thread.}
}
\value{
A list with the result of the function on each item in x.
}
\description{
Apply a function to a list using the cluster
}
\details{
The function will be executed on each element of x in the threads of the cluster. If there are more
elements than threads, the elements will be queued. The progress bar will show the number of
elements that have been completed.
}

