% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Args.R
\name{convertArgsToList}
\alias{convertArgsToList}
\title{Convert arguments used in call to a list}
\usage{
convertArgsToList(matchCall, resultClass = "list")
}
\arguments{
\item{matchCall}{The result of \code{match.call()}.}

\item{resultClass}{The class of the resulting object.}
}
\value{
An object of the class specified in \code{resultClass}.
}
\description{
Convert arguments used in call to a list
}
\details{
Takes the argument values (both default and user-specified) and store them in a list.
}
\examples{
myFun <- function(x = 1, y = 2) {
  return(convertArgsToList(match.call()))
}
}

