% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FakeDf.R
\name{convertFakeDfToDataFrame}
\alias{convertFakeDfToDataFrame}
\title{Convert a fakeDf object to a data.frame}
\usage{
convertFakeDfToDataFrame(fakeDf)
}
\arguments{
\item{fakeDf}{An object of type \code{fakeDf}.}
}
\value{
A data.frame.
}
\description{
Convert a fakeDf object to a data.frame
}

