% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{createConceptSetWorkbook}
\alias{createConceptSetWorkbook}
\title{Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook}
\usage{
createConceptSetWorkbook(fileName, workFolder = NULL, baseUrl,
  included = FALSE, mapped = FALSE)
}
\arguments{
\item{fileName}{Name of a CSV file in the inst/settings folder of the package specifying the
concept sets to add to a workbook. See details for the expected file format.}

\item{workFolder}{Directory location of where the workbook will be saved, defaults to working
directory.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{included}{Should included concepts be included in the workbook?}

\item{mapped}{Should mapped concepts be included in the workbook?}
}
\value{
A xlsx workbook (conceptSetExpressions.xlsx) that includes a list of all concept set IDs and names
and a worksheet for the concepts in each set. Options to include an included concepts and mapped
concepts worksheet for each concept set are avaialble.
}
\description{
Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook
}
\details{
The CSV file should have at least the following fields: \describe{ \item{conceptSetId}{The concept set 
ID in ATLAS.} \item{conceptSetName}{The concept set name that corresponds to the conceptSetId (this 
column can be named anything)} }
}
