% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging.R
\name{createConsoleAppender}
\alias{createConsoleAppender}
\title{Create console appender}
\usage{
createConsoleAppender(layout = layoutSimple)
}
\arguments{
\item{layout}{The layout to be used by the appender.}
}
\description{
Create console appender
}
\details{
Creates an appender that will write to the console.

Deprecated. This function has moved to ParallelLogger.
}
