% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsSchema.R
\name{createResultsSchemaStub}
\alias{createResultsSchemaStub}
\title{Create a results schema stub}
\usage{
createResultsSchemaStub(
  folder,
  outputFile = file.path(folder, "resultsDataModelSpecification.csv")
)
}
\arguments{
\item{folder}{The export folder containing example CSV files.}

\item{outputFile}{The name of the CSV file to be created.}
}
\value{
Invisibly returns the results schema. Is usually executed for the side-effect
of creating the results schema CSV file.
}
\description{
Create a results schema stub
}
\details{
Scans an export folder for CSV files, and created an initial results schema CSV
file as used in Strategus modules. This schema CVS must be manually edited for 
correctness, but should make a good start.
}
