% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{ensurePackageInstalled}
\alias{ensurePackageInstalled}
\title{Ensures a package is installed.}
\usage{
ensurePackageInstalled(pkg)
}
\arguments{
\item{pkg}{The package name to ensure is installed}
}
\description{
Ensures a package is installed.
}
\details{
Checks to see if a package is installed. If it is not, this function
will attempt to install it.

Borrowed and adapted from devtools:
https://github.com/hadley/devtools/blob/ba7a5a4abd8258c52cb156e7b26bb4bf47a79f0b/R/utils.r#L74
}
\examples{
\dontrun{
# This will ensure ggplot is installed:

ensurePackageInstalled(pkg = "ggplot")
                   
}
                       
}
