% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FakeDf.R
\name{fakeDfToFfdf}
\alias{fakeDfToFfdf}
\title{Convert a fakeDf object to ffdf}
\usage{
fakeDfToFfdf(fakeDf)
}
\arguments{
\item{fakeDf}{An object of type \code{fakeDf}.}
}
\value{
An ffdf object.
}
\description{
Convert a fakeDf object to ffdf
}

