% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeFormatting.R
\name{fixHadesLogo}
\alias{fixHadesLogo}
\title{Fix HADES logo in pkgdown output}
\usage{
fixHadesLogo(path = ".")
}
\arguments{
\item{path}{Path to the root of the package for which the pkgdown output needs to be fixed.}
}
\value{
This function returns nothing.
}
\description{
In all HTML files in the docs folder, each occurrence of 'hadesLogo' is replaced with an
HTML image tag referring to the HADES logo.
}
