% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{fixHadesPackagesInLockFile}
\alias{fixHadesPackagesInLockFile}
\title{Fix HADES packages in lock file}
\usage{
fixHadesPackagesInLockFile(
  lockFile = "renv.lock",
  ohdsiGitHubPackages = getOhdsiGitHubPackages()
)
}
\arguments{
\item{lockFile}{The name of the lock file to fix.}

\item{ohdsiGitHubPackages}{Names of R packages that need to be installed from the OHDSI GitHub.}
}
\value{
Returns nothing. Fixes the lock file.
}
\description{
For all HADES packages that must be installed from GitHub, ensures the lock file points to the 
git tag corresponding to the installed version. If no corresponding tag exists, a warning is raised.
}
