% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CodeFormatting.R
\name{formatRFile}
\alias{formatRFile}
\title{Format an R file}
\usage{
formatRFile(file, width.cutoff = 100)
}
\arguments{
\item{file}{The path to the file.}

\item{width.cutoff}{Number of characters that each line should be limited to.}

\item{width}{}
}
\description{
Format an R file
}

