% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CodeFormatting.R
\name{formatRFolder}
\alias{formatRFolder}
\title{Format all R files in a folder}
\usage{
formatRFolder(path = ".", recursive = TRUE, ...)
}
\arguments{
\item{path}{Path to the folder containing the files to format. Only files with the .R
extension will be formatted}

\item{recursive}{Include all subfolders?}
}
\description{
Format all R files in a folder
}

