% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeFormatting.R
\name{formatRText}
\alias{formatRText}
\title{Format R code}
\usage{
formatRText(text, width.cutoff = 100)
}
\arguments{
\item{text}{A character vector with the R code to be formatted.}

\item{width.cutoff}{Number of characters that each line should be limited to.}
}
\value{
A character vector with formatted R code.
}
\description{
Format R code
}
