% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{getCohortInclusionRules}
\alias{getCohortInclusionRules}
\title{Get cohort inclusion rules}
\usage{
getCohortInclusionRules()
}
\value{
A data frame with the names of all inclusion rules.
}
\description{
Get cohort inclusion rules
}
\details{
Parses all cohort definition JSON files in the inst/cohorts folder and extracts the name
of all inclusion rules.
}
