% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{getConceptSetConcepts}
\alias{getConceptSetConcepts}
\title{Get Concept Set Concept Ids}
\usage{
getConceptSetConcepts(baseUrl, setId, vocabSourceKey = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance}

\item{setId}{The concept set id in Atlas}
}
\value{
A list of concept Ids
}
\description{
Get Concept Set Concept Ids
}
\details{
Obtains the full list of concept Ids in a concept set
}
