% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{getCorePackages}
\alias{getCorePackages}
\title{Get a list of R core packages}
\usage{
getCorePackages()
}
\value{
A character vector.
}
\description{
Get a list of R core packages
}
\details{
Returns names of packages that are part of the R code, and can therefore not be installed.
}
