% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{getOhdsiGitHubPackages}
\alias{getOhdsiGitHubPackages}
\title{Get a list of packages in the OHDSI GitHub.}
\usage{
getOhdsiGitHubPackages()
}
\value{
A character vector.
}
\description{
Get a list of packages in the OHDSI GitHub.
}
\details{
Returns names of packages that need to be installed from https://github.com/ohdsi.
}
