% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{insertCirceDefinitionInPackage}
\alias{insertCirceDefinitionInPackage}
\title{Load a Circe definition and insert it into this package}
\usage{
insertCirceDefinitionInPackage(definitionId, name = NULL,
  baseUrl = "http://hix.jnj.com:8080/WebAPI")
}
\arguments{
\item{definitionId}{The number indicating which Circe definition to fetch.}

\item{name}{The name that will be used for the json and SQL files. If not provided, the
name in Circe will be used, but this may not lead to valid file names.}

\item{baseUrl}{The base URL for the WebApi instance.}
}
\description{
Load a Circe definition and insert it into this package
}
\details{
Deprecated. Use \code{\link{insertCohortDefinitionInPackage}} instead.
}
