% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{insertCohortDefinitionInPackage}
\alias{insertCohortDefinitionInPackage}
\title{Load a cohort definition and insert it into this package}
\usage{
insertCohortDefinitionInPackage(definitionId, name = NULL,
  baseUrl = "http://hix.jnj.com:8080/WebAPI")
}
\arguments{
\item{definitionId}{The number indicating which cohort definition to fetch.}

\item{name}{The name that will be used for the json and SQL files. If not provided, the
name in cohort will be used, but this may not lead to valid file names.}

\item{baseUrl}{The base URL for the WebApi instance.}
}
\description{
Load a cohort definition and insert it into this package
}
\details{
Load a cohort definition from a WebApi instance and insert it into this package. This will fetch the
json object and store it in the 'inst/cohorts' folder, and fetch the template SQL and store it in the
'inst/sql/sql_server' folder. Both folders will be created if they don't exist.
}
\examples{
\dontrun{
# This will create 'inst/cohorts/Angioedema.json' and 'inst/sql/sql_server/Angioedema.sql':

insertCohortDefinitionInPackage(282, "Angioedema")
}

}

