% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{insertSqlForCohortTableInPackage}
\alias{insertSqlForCohortTableInPackage}
\title{Insert SQL for creating a cohort table in the package}
\usage{
insertSqlForCohortTableInPackage(statsTables = FALSE)
}
\arguments{
\item{statsTables}{If TRUE, the SQL will also create the four tables needed
when computing inclusion statistics.}
}
\description{
Insert SQL for creating a cohort table in the package
}
\details{
Creates a SQL file called inst/sql/sql_server/CreateCohortTable.sql that
will create an empty cohort table.
}
