% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{isPackageInstalled}
\alias{isPackageInstalled}
\title{Checks if a pacakge is installed}
\usage{
isPackageInstalled(pkg, version = 0)
}
\arguments{
\item{pkg}{The package name}

\item{version}{The minimum version of the package required. Default is 0}
}
\value{
TRUE if the package is installed and at or above the version specified.
}
\description{
Checks if a pacakge is installed
}
\details{
Borrowed from devtools:
https://github.com/hadley/devtools/blob/ba7a5a4abd8258c52cb156e7b26bb4bf47a79f0b/R/utils.r#L44
}
\examples{
\dontrun{
# This will check if devtools is installed:

isPackageInstalled(pkg = "devtools")
                   
}
                       
}
