% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging.R
\name{layoutParallel}
\alias{layoutParallel}
\title{Layout for futile.logger for parellel computing}
\usage{
layoutParallel(level, msg, ...)
}
\arguments{
\item{level}{The level of the message (e.g. "INFO")}

\item{msg}{The message to layout.}

\item{...}{Values to be used in the message.}
}
\description{
A layout function to be used with the \code{flog.layout} function in the \code{futile.logger} package. The layout is 
identical to \code{layout.simple}, except that the thread idenftifier is also included.
}
