% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FakeDf.R
\name{nrowOfFakeDf}
\alias{nrowOfFakeDf}
\title{Return number of rows in fakeDf}
\usage{
nrowOfFakeDf(fakeDf)
}
\arguments{
\item{fakeDf}{The fakeDf object}
}
\value{
Number of rows.
}
\description{
Return number of rows in fakeDf
}

