% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CodeChecking.R
\name{ohdsiLintrFile}
\alias{ohdsiLintrFile}
\title{Check R file for code errors}
\usage{
ohdsiLintrFile(file)
}
\arguments{
\item{file}{The R file to check}
}
\description{
Check R file for code errors
}
\details{
This function uses the lintr package to check the code.
}

