% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CodeChecking.R
\name{ohdsiLintrFolder}
\alias{ohdsiLintrFolder}
\title{Check all R files in a folder}
\usage{
ohdsiLintrFolder(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{Path to the folder containing the files to check. Only files with the .R
extension will be checked.}

\item{recursive}{Include all subfolders?}
}
\description{
Check all R files in a folder
}
\details{
This function uses the lintr package to check the code.
}

