% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiRTools.R
\docType{package}
\name{OhdsiRTools-package}
\alias{OhdsiRTools}
\alias{OhdsiRTools-package}
\title{OhdsiRTools: Tools Used by Observational Health Data Science and Informatics (OHDSI)}
\description{
Includes functions to format and check syntax of R code and packages following the 'OHDSI'
    R style guidelines. Support for parallel computation, logging, and function automation. Functionality
    for interacting with instances of the open source 'OHDSI' WebApi, which can be found at <https://github.com/OHDSI/WebAPI>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/OhdsiRTools}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiRTools/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
