% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Args.R
\name{prettyPrint}
\alias{prettyPrint}
\title{Print a list of objects}
\usage{
prettyPrint(object)
}
\arguments{
\item{object}{The list to print.}
}
\description{
Print a list of objects
}
\details{
Will print nested lists using indentation.
}

