% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FakeDf.R
\name{rbindFakeDfs}
\alias{rbindFakeDfs}
\title{Bind two fakeDf objects by rows}
\usage{
rbindFakeDfs(fakeDf1, fakeDf2)
}
\arguments{
\item{fakeDf1}{The first fakeDf object.}

\item{fakeDf2}{The second fakeDf object.}
}
\value{
A fakeDf object with the rows of both fakeDf1 and fakeDf2.
}
\description{
Bind two fakeDf objects by rows
}

