% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{restoreEnvironment}
\alias{restoreEnvironment}
\title{Restore the R environment to a snapshot}
\usage{
restoreEnvironment(snapshot, stopOnWrongRVersion = FALSE)
}
\arguments{
\item{snapshot}{The snapshot data frame as generated using the
\code{\link{takeEnvironmentSnapshot}} function.}

\item{stopOnWrongRVersion}{Should the function stop when the wrong version of R is installed? Else
just a warning will be thrown when the version doesn't match.}
}
\description{
Restore the R environment to a snapshot
}
\details{
This function restores the R environment to a previous snapshot, meaning all the packages will be
restored to the versions they were at at the time of the snapshot.
Note: on Windows you will very likely need to have RTools installed to build the various packages.
}
\examples{
\dontrun{
snapshot <- takeEnvironmentSnapshot("OhdsiRTools")
write.csv(snapshot, "snapshot.csv")

# 5 years later

snapshot <- read.csv("snapshot.csv")
restoreEnvironment(snapshot)
}
}
