% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{restoreEnvironmentFromPackage}
\alias{restoreEnvironmentFromPackage}
\title{Restore environment stored in package}
\usage{
restoreEnvironmentFromPackage(
  pathToCsv = "inst/settings/rEnvironmentSnapshot.csv",
  stopOnWrongRVersion = FALSE,
  strict = FALSE,
  skipLast = TRUE
)
}
\arguments{
\item{pathToCsv}{The path for saving the snapshot (as CSV file).}

\item{stopOnWrongRVersion}{Should the function stop when the wrong version of R is installed? Else
just a warning will be thrown when the version doesn't match.}

\item{strict}{If TRUE, the exact version of each package will installed. If FALSE, a
package will only be installed if (a) a newer version is required than
currently installed, or (b) the major version number is different.}

\item{skipLast}{Skip last entry in snapshot? This is usually the study package that
needs to be installed manually.}
}
\description{
Restore environment stored in package
}
\details{
This function restores all packages (and package versions) described in the environment snapshot
stored in the package currently being developed. The default location is
\code{inst/settings/rEnvironmentSnapshot.csv}.
}
\examples{
\dontrun{
restoreEnvironmentFromPackage()
}

}
