% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{selectFromList}
\alias{selectFromList}
\title{Select variables from a list of objects of the same type}
\usage{
selectFromList(x, select)
}
\arguments{
\item{x}{A list of objects of the same type.}

\item{select}{A character vector of names of variables to select.}
}
\description{
Select variables from a list of objects of the same type
}
