% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FakeDf.R
\name{subsetFakeDf}
\alias{subsetFakeDf}
\title{Create a subset of fakeDf object}
\usage{
subsetFakeDf(fakeDf, index)
}
\arguments{
\item{fakeDf}{The fakeDf object to subset.}

\item{index}{An integer vector identifying the rows to extract.}
}
\value{
An object of type \code{fakeDf}, which is a list of vectors.
}
\description{
Create a subset of fakeDf object
}

