% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FakeDf.R
\name{subsetFfdfToFakeDf}
\alias{subsetFfdfToFakeDf}
\title{Create a subset of an ffdf object as a fakeDf object}
\usage{
subsetFfdfToFakeDf(ffdf, index)
}
\arguments{
\item{ffdf}{The ffdf object to subset.}

\item{index}{An integer vector or ri object identifying the rows to extract.}
}
\value{
An object of type \code{fakeDf}, which is a list of vectors.
}
\description{
Create a subset of an ffdf object as a fakeDf object
}

