% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{takeEnvironmentSnapshot}
\alias{takeEnvironmentSnapshot}
\title{Take a snapshot of the R environment}
\usage{
takeEnvironmentSnapshot(rootPackage)
}
\arguments{
\item{rootPackage}{The name of the root package}
}
\value{
A data frame listing all the dependencies of the root package and their version numbers, in the
order in which they should be installed.
}
\description{
Take a snapshot of the R environment
}
\details{
This function records all versions used in the R environment that are used by one root package.
This can be used for example to restore the environment to the state it was when a particular study
package was run using the \code{\link{restoreEnvironment}} function.
}
\examples{
snapshot <- takeEnvironmentSnapshot("OhdsiRTools")
snapshot

}
