% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyrightYearUpdate.R
\name{updateCopyrightYearFile}
\alias{updateCopyrightYearFile}
\title{Update the copyright year in a R or SQL file}
\usage{
updateCopyrightYearFile(file)
}
\arguments{
\item{file}{The path to the file.}
}
\description{
Update the copyright year in a R or SQL file
}

