% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyrightYearUpdate.R
\name{updateCopyrightYearFolder}
\alias{updateCopyrightYearFolder}
\title{Update the copyright year in all R and SQL files in a folder}
\usage{
updateCopyrightYearFolder(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{Path to the folder containing the files to update. Only files with the .R and .SQL
extension will be updated.}

\item{recursive}{Include all subfolders?}
}
\description{
Update the copyright year in all R and SQL files in a folder
}
\examples{
\dontrun{
updateCopyrightYearFolder()
}
}

