% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyrightYearUpdate.R
\name{updatePackageName}
\alias{updatePackageName}
\title{Update the package name in a R or SQL file}
\usage{
updatePackageName(file, packageName)
}
\arguments{
\item{file}{The path to the file.}

\item{packageName}{The replacement package name}
}
\description{
Update the package name in a R or SQL file
}
