% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyrightYearUpdate.R
\name{updatePackageNameFolder}
\alias{updatePackageNameFolder}
\title{Update the package name in all R and SQL files in a folder}
\usage{
updatePackageNameFolder(path = ".", packageName, recursive = TRUE)
}
\arguments{
\item{path}{Path to the folder containing the files to update. Only files with the .R and
.SQL extension will be updated.}

\item{packageName}{The replacement package name}

\item{recursive}{Include all subfolders?}
}
\description{
Update the package name in all R and SQL files in a folder
}
\examples{
\dontrun{
updateCopyrightYearFolder()
}
}
