% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{createRenvLockFile}
\alias{createRenvLockFile}
\title{Create a renv lock file}
\usage{
createRenvLockFile(
  rootPackage,
  mode = "auto",
  includeRootPackage = TRUE,
  additionalRequiredPackages = NULL,
  ohdsiGitHubPackages = getOhdsiGitHubPackages(),
  ohdsiStudiesGitHubPackages = rootPackage,
  fileName = "renv.lock",
  restart = TRUE
)
}
\arguments{
\item{rootPackage}{The name of the root package, the package that we'd like to be
able to run in the end.}

\item{mode}{Can be "auto" or "description". See details.}

\item{includeRootPackage}{Include the root package in the renv file?}

\item{additionalRequiredPackages}{Additional packages we want to have installed (with their
dependencies), such as 'keyring'. Ignored if \code{mode = "auto"}.}

\item{ohdsiGitHubPackages}{Names of R packages that need to be installed from the OHDSI
GitHub.}

\item{ohdsiStudiesGitHubPackages}{Names of R packages that need to be installed from the
OHDSI-Studies GitHub.}

\item{fileName}{Name of the lock file to be generated. Ignored if \code{mode = "auto"}.}

\item{restart}{Boolean; attempt to restart the R session after initializing the project?}
}
\value{
Does not return a value. Is executed for the side-effect of creating the lock file.
}
\description{
Create a renv lock file
}
\details{
Create a lock file that allows reconstruction of the R environment using the \code{renv} package.
This function will include the root file and all of its dependencies in the lock file, requiring
the same package versions as currently installed on this computer.

If \code{mode = "auto"}, this function will invoke \code{renv::init()}, which in turn will scan the
project folders for any dependencies that are referenced. Afterwards, references to OHDSI packages will
be altered so the correct GitHub tags are used for the installed versions.

If \code{mode = "description"}, this function will assume the project is a full-fledged R package with
up-to-date DESCRIPTION, and will only install the dependencies listed in the DESCRIPION.

The second option tends to lead to smaller lock files, but requires all dependencies are accurately listed
in the DESCRIPTION file of the study package.
}
