% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{createRenvLockFile}
\alias{createRenvLockFile}
\title{Create a renv lock file}
\usage{
createRenvLockFile(
  rootPackage,
  ohdsiGitHubPackages = getOhdsiGitHubPackages(),
  ohdsiStudiesGitHubPackages = rootPackage,
  fileName = "renv.lock"
)
}
\arguments{
\item{rootPackage}{The name of the root package, the package that we'd like to be
able to run in the end.}

\item{ohdsiGitHubPackages}{Names of R packages that need to be installed from the OHDSI
GitHub.}

\item{ohdsiStudiesGitHubPackages}{Names of R packages that need to be installed from the
OHDSI-Studies GitHub.}

\item{fileName}{Name of the lock file to be generated.}
}
\value{
Does not return a value. Is executed for the side-effect of creating the lock file.
}
\description{
Create a renv lock file
}
\details{
Create a lock file that allows recontruction of the R environment using the \code{renv} package.
This function will include the root file and all of its dependencies in the lock file, requiring
the same package versions as currently installed on this computer.
}
