% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeChecking.R
\name{findNonAsciiStringsInFolder}
\alias{findNonAsciiStringsInFolder}
\title{Find non-ASCII strings in files}
\usage{
findNonAsciiStringsInFolder(path = ".", recursive = TRUE, pattern = "*.R$")
}
\arguments{
\item{path}{Path to the folder containing the files matching the pattern parameter.}

\item{recursive}{If TRUE, subfolders will also be searched for files matching the pattern parameter.}

\item{pattern}{The regular expression to use for selecting files. The default is .R files.}
}
\value{
A table listing the lines per file containing non-ASCII characters.
}
\description{
Find non-ASCII string in files
}
