% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeFormatting.R
\name{formatRFile}
\alias{formatRFile}
\title{Format an R file}
\usage{
formatRFile(file, width.cutoff = 100)
}
\arguments{
\item{file}{The path to the file.}

\item{width.cutoff}{Number of characters that each line should be limited to.}
}
\description{
Format an R file
}
\details{
DEPRECRATED. Please use \code{styler::style_file} instead.
}
