% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeFormatting.R
\name{formatRFolder}
\alias{formatRFolder}
\title{Format all R files in a folder}
\usage{
formatRFolder(path = ".", recursive = TRUE, skipAutogenerated = TRUE,
  ...)
}
\arguments{
\item{path}{Path to the folder containing the files to format. Only files with the .R
extension will be formatted.}

\item{recursive}{Include all subfolders?}

\item{skipAutogenerated}{Skip auto-generated files such as RcppExports.R?}

\item{...}{Parameters to be passed on the the formatRFile function}
}
\description{
Format all R files in a folder
}
\examples{
\donttest{
formatRFolder()
}
}
