% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WebApiTools.R
\name{insertCirceDefinitionInPackage}
\alias{insertCirceDefinitionInPackage}
\title{Load a Circe definition and insert it into this package}
\usage{
insertCirceDefinitionInPackage(definitionId, name = NULL,
  baseUrl = "http://hixbeta.jnj.com:8081/WebAPI")
}
\arguments{
\item{name}{The name that will be used for the json and SQL files. If not provided, the
name in Circe will be used, but this may not lead to valid file names.}

\item{baseUrl}{The base URL for the WebApi instance.}

\item{defintionId}{The number indicating which Circe definition to fetch.}
}
\description{
Load a Circe definition and insert it into this package
}
\details{
Load a Circe definition from a WebApi instance and insert it into this package. This will
fetch the json object and store it in the 'inst/circe' folder, and fetch the template SQL
and store it in the 'inst/sql/sql_server' folder. Both folders will be created if they
don't exist.
}
\examples{
\dontrun{
# This will create 'inst/circe/MyocardialInfarction.json' and
# 'inst/sql/sql_server/MyocardialInfarction.sql':

insertCirceDefinitionInPackage(280, "MyocardialInfarction")
}
}

