% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSnapshot.R
\name{insertEnvironmentSnapshotInPackage}
\alias{insertEnvironmentSnapshotInPackage}
\title{Store snapshot of the R environment in the package}
\usage{
insertEnvironmentSnapshotInPackage(
  rootPackage,
  pathToCsv = "inst/settings/rEnvironmentSnapshot.csv"
)
}
\arguments{
\item{rootPackage}{The name of the root package}

\item{pathToCsv}{The path for saving the snapshot (as CSV file).}
}
\description{
Store snapshot of the R environment in the package
}
\details{
This function records all versions used in the R environment that are used by one root package, and
stores them in a CSV file in the R package that is currently being developed. The default location
is \code{inst/settings/rEnvironmentSnapshot.csv}.This can be used for example to restore the
environment to the state it was when a particular study package was run using the
\code{\link{restoreEnvironment}} function.
}
\examples{
\dontrun{
insertEnvironmentSnapshotInPackage("OhdsiRTools")
}

}
