% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyrightYearUpdate.R
\name{updateCopyrightYearFile}
\alias{updateCopyrightYearFile}
\title{Update the copyright year in an R file}
\usage{
updateCopyrightYearFile(file, width.cutoff = 100)
}
\arguments{
\item{file}{The path to the file.}

\item{width.cutoff}{Number of characters that each line should be limited to.}
}
\description{
Update the copyright year in an R file
}

